+++ library/std/src/os/unix/process.rs
@@ -593,5 +593,5 @@ impl From<OwnedFd> for process::ChildStderr {
 #[must_use]
 #[stable(feature = "unix_ppid", since = "1.27.0")]
 pub fn parent_id() -> u32 {
-    crate::sys::os::getppid()
+    crate::sys::process::getppid()
 }
