+++ library/std/src/sys/pal/motor/os.rs
@@ -62,7 +62,3 @@ pub fn temp_dir() -> PathBuf {
 pub fn home_dir() -> Option<PathBuf> {
     None
 }
-
-pub fn getpid() -> u32 {
-    panic!("Pids on Motor OS are u64.")
-}
