+++ library/std/src/sys/pal/xous/os.rs
@@ -118,7 +118,3 @@ pub fn temp_dir() -> PathBuf {
 pub fn home_dir() -> Option<PathBuf> {
     None
 }
-
-pub fn getpid() -> u32 {
-    panic!("no pids on this platform")
-}
