+++ library/std/src/sys/process/mod.rs
@@ -27,9 +27,11 @@ cfg_select! {
 mod env;
 
 pub use env::CommandEnvs;
+#[cfg(target_family = "unix")]
+pub use imp::getppid;
 pub use imp::{
     ChildPipe, Command, CommandArgs, EnvKey, ExitCode, ExitStatus, ExitStatusError, Process, Stdio,
-    read_output,
+    getpid, read_output,
 };
 
 #[cfg(any(
