+++ library/std/src/sys/process/unix/common.rs
@@ -679,3 +679,11 @@ pub fn read_output(
         }
     }
 }
+
+pub fn getpid() -> u32 {
+    unsafe { libc::getpid() as u32 }
+}
+
+pub fn getppid() -> u32 {
+    unsafe { libc::getppid() as u32 }
+}
