+++ library/std/src/sys/process/unix/mod.rs
@@ -23,5 +23,7 @@ cfg_select! {
 
 pub use imp::{ExitStatus, ExitStatusError, Process};
 
-pub use self::common::{ChildPipe, Command, CommandArgs, ExitCode, Stdio, read_output};
+pub use self::common::{
+    ChildPipe, Command, CommandArgs, ExitCode, Stdio, getpid, getppid, read_output,
+};
 pub use crate::ffi::OsString as EnvKey;
